pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "OrbFIX-Testbench"
version = "0.1.0"
description = "OrbFIX bench command utilities and test runners"
readme = "README.md"
requires-python = ">=3.8.1"
dependencies = [
  "pyserial",
  "typer",
  "pydantic",
  "PyYAML",
  "importlib_resources>=5.12; python_version < '3.9'",
  "pycrc",
  "pyusb",
  "windows-curses"
]

[project.scripts]
orbfix = "orbfix.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["orbfix*"]

[tool.setuptools.package-data]
orbfix = ["scripts/*.sh", "scripts/utils/*.sh"]
